add new styles to my components
