+++ src/App.css
@@ -1,52 +1,8 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
 .app-wrapper{
-  display: grid;
-  width: 1200px;
-  margin: 0 auto;
 
-  grid-template-areas:
-  "h h"
-  "n c";
-  grid-auto-rows: 60px 1fr;
-    grid-auto-columns: 2fr 10fr;
-    /*grid-gap: 10px;*/
+}
+.appContent{
+  display: flex;
 }
