+++ src/App.tsx
@@ -1,15 +1,18 @@
 import React from 'react';
 import './App.css';
-import Header from './components/Header';
-import Navbar from './components/Navbar';
-import Profile from './components/Profile';
+import Header from './components/Header/Header';
+import Navbar from './components/Navbar/Navbar';
+import Profile from './components/Profile/Profile';
 
 const App = () => {
   return (
-    <div className = "app-wrapper">
+    <div className = "appWrapper">
         <Header />
-        <Navbar />
-        <Profile title={'Make love not war'}/>
+        <div className="appContent">
+            <Navbar />
+            <Profile title={'Make love not war'}/>
+        </div>
+
 
     </div>
   );
