--- src/components/Header.module.css
@@ -1,4 +1,0 @@
-.header{
-    grid-area: h;
-    background-color: green;
-}
