--- src/components/Navbar.module.css
@@ -1,10 +1,0 @@
-.nav{
-    grid-area: n;
-    background-color: #61dafb;
-}
-.item{
-    color: white;
-}
-.item.active{
-    color: gold;
-}
