+++ src/components/Navbar/Navbar.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import classes from './Navbar.module.css';
+
 console.log(classes);
 
 const Navbar = () => {
@@ -8,7 +9,7 @@ const Navbar = () => {
             <div className={`${classes.item} ${classes.active}`}>
                 <a>Profile</a>
             </div>
-            <div className={classes.item} >
+            <div className={classes.item}>
                 <a>Messages</a>
             </div>
             <div className={classes.item}>
