--- src/components/Profile.module.css
@@ -1,11 +1,0 @@
-.content{
-    grid-area: c;
-    background-color: burlywood;
-}
-.content img{
-    width: 1000px;
-    height: 600px;
-}
-.item{
-    color: greenyellow;
-}
