+++ src/components/Profile/Profile.module.css
@@ -1,0 +1,17 @@
+.content{
+    width: 100%;
+    margin-top:5px;
+    padding-bottom:20px;
+
+    background-color: #fce5c7;
+}
+.content img{
+    width: 100%;
+    height: 300px;
+}
+.item{
+    color: #f50808;
+}
+.postsSection{
+    padding-left: 40px;
+}
